+++ packages/admin-ui/src/index.ts
@@ -1,5 +1,4 @@
 export { default as Breadcrumbs } from './breadcrumbs';
-export { default as Navigation } from './navigation';
 export { default as NavigableRegion } from './navigable-region';
 export { default as Page } from './page';
 export { getAdminThemeColors } from './admin-theme-colors';